+++ src/rust/src/lib.rs
@@ -1,6 +1,5 @@
 use extendr_api::prelude::*;
 
-mod laser_frame;
 mod epidemic;
 mod distributions;
 mod migration;
@@ -13,13 +12,11 @@ mod kmestimator;
 mod mortality;
 mod measles;
 mod births;
-pub use laser_frame::LaserFrame;
 pub use distributions::Distribution;
 pub use column::Column;
 
 extendr_module! {
     mod razer;
-    use laser_frame;
     use epidemic;
     use distributions;
     use migration;